fix(conflict): fix conflicting files

+++ controllers/FollowController.js
@@ -21,9 +21,7 @@ class FollowController {
       return res.status(404).json({ status: 404, message: 'User not found' });
     }
 
-    await Follow.findOrCreate({
-      where: { followee: userId, follower: currentUser.id }
-    });
+    await Follow.findOrCreate({ where: { followee: userId, follower: currentUser.id } });
 
     return res.status(201).json({ status: '201', message: `You followed ${followee.firstName}` });
   }
@@ -48,9 +46,7 @@ class FollowController {
       return res.status(404).json({ status: 404, message: 'User not found' });
     }
 
-    await Follow.destroy({
-      where: { followee: userId, follower: currentUser.id }
-    });
+    await Follow.destroy({ where: { followee: userId, follower: currentUser.id } });
 
     return res
       .status(200)
